passphrase: Move passphrase-related bits to a separate widget
Issue #290
